requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
xgboost==1.7.6
seaborn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
streamlit
streamlit-autorefresh
plotly-express
-e .